SCP11: Fix data types for big endian

+++ aes_cmac/aes_cmac.h
@@ -48,7 +48,7 @@ typedef struct {
 int YH_INTERNAL aes_cmac_init(aes_context *aes_ctx, aes_cmac_context_t *ctx);
 int YH_INTERNAL aes_cmac_encrypt(aes_cmac_context_t *ctx,
                                  const uint8_t *message,
-                                 const uint16_t message_len, uint8_t *mac);
+                                 const uint32_t message_len, uint8_t *mac);
 void YH_INTERNAL aes_cmac_destroy(aes_cmac_context_t *ctx);
 
 #endif //YUBICO_PIV_TOOL_AES_CMAC_H
